sih_requirements.txt
Flask==2.3.2
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
cachetools==5.3.1
groq==0.9.0
PyPDF2==3.0.1
python-docx==0.8.11
Flask
flask-cors
pillow>=10.0.0,<11.0.0
langchain
langchain-groq
cassandra-driver
cassio
python-dotenv
langchain-core
langchain-nvidia-ai-endpoints==0.1.1
langchain-community
llama-index
llama-index-llms-groq
python-dotenv
bs4
qdrant-client